requirements.txt
python-telegram-bot==20.6
pymongo==4.5.0
python-dotenv==1.0.0
dnspython==2.4.2 